Add PaymentStatusService
